JsonMoveGroup: new toplevel output for move generators
This groups multiple JsonMoves under a single entry.  The intent is
to reduce the total number of patch applications by grouping together
changes that are extremely unlikely to need to be split apart into
separate entries.  Since the algorithm for determining how to apply
changes to the configuration is recursive and exponential, reducing
the total number that might need to be applied should greatly reduce
the overall time required.