BulkKeyGroupLowLevelMoveGenerator: Group updates across keys for the same table
This is a Wrapper around BulkLowLevelMoveGenerator that groups the leaf
operations together spanning multiple table keys.  For example if someone
wants to update PORT/Ethernet0/description and PORT/Ethernet1/description
at the same time, it is safe to do so in the same patch group. This grouping
exists in order to attempt to optimize the fast path when there are a lot
of changes to the same table and there are often very few
cross-dependencies.  This will bring down the patch set count considerably
for a lot of change operations.  We do still fall back to other more
primitive generators if the validators fail so this is an optimization that
otherwise doesn't have any impact on the overall outcome, only performance.